+++ HorseJs/WindowBase.h
@@ -6,6 +6,7 @@ public:
 	WindowBase();
 	WindowBase(const WindowBase&) = delete;
 	WindowBase& operator=(const WindowBase&) = delete;
+	static LRESULT CALLBACK RouteWindowMessage(HWND hWnd, UINT msg, WPARAM wParam, LPARAM lParam);
 	LRESULT CALLBACK WindowProc(HWND hwnd, UINT msg, WPARAM wparam, LPARAM lparam);
 	LRESULT HitTest(HWND hwnd, POINT cursor);
 	HWND hWnd;