+++ HorseJs/WindowNative.h
@@ -7,12 +7,14 @@
 #include <wil/com.h>
 #include "WebView2.h"
 #include "WindowBase.h"
+
 class WindowNative:public WindowBase
 {
 public:
 	WindowNative();
 	WindowNative(const WindowNative&) = delete;
 	WindowNative& operator=(const WindowNative&) = delete;	
+	
 private:
 	HRESULT CreateBrowserView(HRESULT result, ICoreWebView2Controller* controller);
 	HRESULT NavigationStarting(ICoreWebView2* webview, ICoreWebView2NavigationStartingEventArgs* args);